refactor: move model/dto -> application/dto
